Bump to 1.58.2 and add Windows support
By switching to the Meson build system, it becomes tractable to build on
Windows. The price is that Meson only works on Python >= 3.6, so we need to
drop support for older Pythons even though gobject-introspection nominally
supports them. In my judgment this is OK.

We still need a few patches to get everything working on all three platforms,
but nothing too invasive.